debug du init_bss()

+++ td/init.c
@@ -1,8 +1,8 @@
 #include <stdint.h>
 
-extern uint16_t _etext, _data, _edata, _rodata, _erodata, _bstart, _bend;
+extern uint8_t _etext, _data, _edata, _rodata, _erodata, _bstart, _bend;
 
-uint16_t *dst;
+uint8_t *dst;
 
 void init_bss() {
     for (dst = &_bstart; dst < &_bend; dst++)
